+++ ovp_users/models.py
@@ -68,5 +68,6 @@ class PasswordRecoveryToken(models.Model):
   def save(self, *args, **kwargs):
     if not self.pk:
       self.token = uuid.uuid4()
+      self.user.mailing().sendRecoveryToken({'token': self.token})
 
     super(PasswordRecoveryToken, self).save(*args, **kwargs)